+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Added blaze slayer phase display
 + Added toggle for blaze slayer colored mobs
 + Mark the right dagger to use for blaze slayer in the dagger overlay.
++ Added option to select the first, left sided dagger for the display.
 
 ### Stats Tuning
 + Show the tuning stats in the Thaumaturgy inventory.